+++ app/helpers/form_helper.rb
@@ -1,0 +1,6 @@
+module FormHelper
+  def list_of_organisations
+    @list_of_organisations ||=
+      [OpenStruct.new(name: "", id: ""), Organisation.all].flatten
+  end
+end