+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_12_11_114801) do
+ActiveRecord::Schema.define(version: 2019_12_13_123607) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -73,7 +73,11 @@ ActiveRecord::Schema.define(version: 2019_12_11_114801) do
     t.uuid "fund_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.uuid "provider_id"
+    t.uuid "receiver_id"
     t.index ["fund_id"], name: "index_transactions_on_fund_id"
+    t.index ["provider_id"], name: "index_transactions_on_provider_id"
+    t.index ["receiver_id"], name: "index_transactions_on_receiver_id"
   end
 
   create_table "users", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|